Added random hash for output_file and fix the deleteFile.py script

+++ runScript.py
@@ -6,24 +6,28 @@ from deleteFiles import DeleteFiles
 
 
 class RunScript():
-    def __init__(self, output_file):
-        self.output_file = output_file
 
     def run_script(self):
-        args = GetArguments().get_args()    
+        arg_parser = GetArguments()
+        check_if_report_path_exist = arg_parser.check_if_report_path_exist()
+        args = arg_parser.get_args()
+        output_file = arg_parser.get_output_file()
+
         try:
-            gitleaks_runner = RunGitleaks(args, self.output_file)
+            gitleaks_runner = RunGitleaks(args, output_file)
             gitleaks_runner.run_tool()
             gitleaks_runner.check_result()
         except Exception as e:
             print(f"Error: {e}")
 
-        processor = ProcessData(self.output_file)
+        processor = ProcessData(output_file)
         try:
             structured_findings = processor.process_data()
+            processor.write_findings_to_file(structured_findings)
             printer = LogData(structured_findings)
             printer.print_to_console()
-            deleter = DeleteFiles(self.output_file)
-            deleter.delete_file()
+            if check_if_report_path_exist is False:
+                deleter = DeleteFiles(output_file,args)
+                deleter.delete_file()
         except Exception as e:
             print(f"An error occurred: {e}")
